Realizada a implementação da política de acesso as formas de pagamento.

+++ app/Http/Controllers/FormPaymentsController.php
@@ -15,22 +15,26 @@ class FormPaymentsController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', FormPayment::class);
         return FormPayment::create($request->all());
     }
 
     public function show(FormPayment $form_payment)
     {
+        $this->authorize('view', $form_payment);
         return $form_payment;
     }
 
     public function update(Request $request, FormPayment $form_payment)
     {
+        $this->authorize('update', $form_payment);
         $form_payment->update($request->all());
         return $form_payment;
     }
 
     public function destroy(FormPayment $form_payment)
     {
+        $this->authorize('delete', $form_payment);
         $form_payment->delete();
         return $form_payment;
     }
